refactor: logging 위해 폴더 구조 변경
