+++ src/main/java/com/burntoburn/easyshift/service/login/UserService.java
@@ -5,9 +5,8 @@ import com.burntoburn.easyshift.entity.user.User;
 import com.burntoburn.easyshift.repository.user.UserRepository;
 import com.burntoburn.easyshift.service.login.dto.LoginRequest;
 import com.burntoburn.easyshift.service.login.dto.SignUpRequest;
-import com.burntoburn.easyshift.session.SessionService;
+import com.burntoburn.easyshift.service.session.SessionService;
 import jakarta.servlet.http.HttpSession;
-import java.util.Map;
 import java.util.NoSuchElementException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
